+++ src/main/java/com/uid2/admin/job/jobsync/client/ClientKeySyncJob.java
@@ -1,35 +1,36 @@
-package com.uid2.admin.job.jobsync;
+package com.uid2.admin.job.jobsync.client;
 
 import com.uid2.admin.job.model.Job;
 import com.uid2.admin.model.PrivateSiteDataMap;
-import com.uid2.admin.model.Site;
 import com.uid2.admin.store.MultiScopeStoreWriter;
 import com.uid2.admin.util.PrivateSiteUtil;
+import com.uid2.shared.auth.ClientKey;
 import com.uid2.shared.auth.OperatorKey;
 
 import java.util.Collection;
 
-public class SiteSyncJob implements Job {
+public class ClientKeySyncJob implements Job {
     private final Collection<OperatorKey> globalOperators;
-    private final Collection<Site> globalSites;
-    private final MultiScopeStoreWriter<Collection<Site>> multiScopeStoreWriter;
+    private final Collection<ClientKey> globalClientKeys;
+    private final MultiScopeStoreWriter<Collection<ClientKey>> multiScopeStoreWriter;
 
-    public SiteSyncJob(
-            MultiScopeStoreWriter<Collection<Site>> multiScopeStoreWriter, Collection<Site> globalSites,
+    public ClientKeySyncJob(
+            MultiScopeStoreWriter<Collection<ClientKey>> multiScopeStoreWriter,
+            Collection<ClientKey> globalClientKeys,
             Collection<OperatorKey> globalOperators) {
-        this.globalSites = globalSites;
+        this.globalClientKeys = globalClientKeys;
         this.globalOperators = globalOperators;
         this.multiScopeStoreWriter = multiScopeStoreWriter;
     }
 
     @Override
     public String getId() {
-        return "global-to-site-scope-sync-sites";
+        return "global-to-site-scope-sync-clientKeys";
     }
 
     @Override
     public void execute() throws Exception {
-        PrivateSiteDataMap<Site> desiredState = PrivateSiteUtil.getSites(globalSites, globalOperators);
+        PrivateSiteDataMap<ClientKey> desiredState = PrivateSiteUtil.getClientKeys(globalOperators, globalClientKeys);
         multiScopeStoreWriter.uploadIfChanged(desiredState, null);
     }
 }
