+++ src/main/java/com/uid2/admin/job/jobsync/site/SiteSyncJob.java
@@ -1,36 +1,35 @@
-package com.uid2.admin.job.jobsync;
+package com.uid2.admin.job.jobsync.site;
 
 import com.uid2.admin.job.model.Job;
 import com.uid2.admin.model.PrivateSiteDataMap;
+import com.uid2.admin.model.Site;
 import com.uid2.admin.store.MultiScopeStoreWriter;
 import com.uid2.admin.util.PrivateSiteUtil;
-import com.uid2.shared.auth.ClientKey;
 import com.uid2.shared.auth.OperatorKey;
 
 import java.util.Collection;
 
-public class ClientKeySyncJob implements Job {
+public class SiteSyncJob implements Job {
     private final Collection<OperatorKey> globalOperators;
-    private final Collection<ClientKey> globalClientKeys;
-    private final MultiScopeStoreWriter<Collection<ClientKey>> multiScopeStoreWriter;
+    private final Collection<Site> globalSites;
+    private final MultiScopeStoreWriter<Collection<Site>> multiScopeStoreWriter;
 
-    public ClientKeySyncJob(
-            MultiScopeStoreWriter<Collection<ClientKey>> multiScopeStoreWriter,
-            Collection<ClientKey> globalClientKeys,
+    public SiteSyncJob(
+            MultiScopeStoreWriter<Collection<Site>> multiScopeStoreWriter, Collection<Site> globalSites,
             Collection<OperatorKey> globalOperators) {
-        this.globalClientKeys = globalClientKeys;
+        this.globalSites = globalSites;
         this.globalOperators = globalOperators;
         this.multiScopeStoreWriter = multiScopeStoreWriter;
     }
 
     @Override
     public String getId() {
-        return "global-to-site-scope-sync-clientKeys";
+        return "global-to-site-scope-sync-sites";
     }
 
     @Override
     public void execute() throws Exception {
-        PrivateSiteDataMap<ClientKey> desiredState = PrivateSiteUtil.getClientKeys(globalOperators, globalClientKeys);
+        PrivateSiteDataMap<Site> desiredState = PrivateSiteUtil.getSites(globalSites, globalOperators);
         multiScopeStoreWriter.uploadIfChanged(desiredState, null);
     }
 }
